requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.18.0
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.7.0
fsspec==2025.7.0
gotrue==2.12.3
gunicorn==22.0.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
httpx-sse==0.4.1
huggingface-hub==0.34.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-anthropic==0.3.18
langchain-community==0.3.27
langchain-core==0.3.72
langchain-huggingface==0.0.3
langchain-text-splitters==0.3.9
langsmith==0.4.10
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.2.1
numpy==1.26.4
orjson==3.11.1
packaging==24.2
pgvector==0.2.0
pillow==11.3.0
postgrest==0.16.11
propcache==0.3.2
psycopg2-binary==2.9.9
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pypdf==5.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
realtime==1.0.6
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==2.7.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
storage3==0.7.7
StrEnum==0.4.15
supabase==2.5.0
supafunc==0.4.7
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.54.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websockets==12.0
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
